throttles api to archive 100k+ messages
I found that I needed to throttle API requests when archiving rooms with more than a year or so of messages. The app would silently die after receiving ~10,000 - 30,000 records. With the delay I was able to archive 100k+ messages from a single room. 

Note: experiment with the delay if it is too conservative for you.